requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
idna==2.10
numpy==1.20.2
palladium-python==0.1.5
pandas==1.2.4
pymongo==3.11.4
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
viper-python==0.1.3
